added very basic editor
added the barebones of the editor, not at all finished with it and I still need to add functionality. Also added the start functions of the flashcards and generate quiz buttons on the study set template. Still have a lot of work to do on all of it but its a start.